playground_requirements.txt
# SnowLeopard Playground requirements
snowleopard
python-dotenv
